+++ lib/src/queries.dart
@@ -4,6 +4,38 @@
 // All rights reserved. Use of this source code is governed by the
 // MIT license that can be found in the LICENSE file.
 
+import 'dart:convert';
+import 'package:graphql/client.dart';
+
+dynamic decodeToken(String token) {
+  if(token == null) {
+    //TODO: throw unexpected parse exception: user token is null, cannot decode it
+    return null;
+  }
+  final parts = token.split('.');
+  if (parts.length != 3) {
+    //TODO: throw unexpected parse exception: user token is not in the right format: cannot decode it. Are you sure you are connected to Krypton Auth?
+    return null;
+  }
+  final payload = parts[1];
+  var normalized = base64Url.normalize(payload);
+  var resp = utf8.decode(base64Url.decode(normalized));
+  return json.decode(resp);
+}
+
+
+Future<QueryResult> query(QueryEnum queryEnum,
+    Map<String, dynamic> variables, GraphQLClient graphQLClient) async {
+  final QueryOptions _options = QueryOptions(
+    documentNode: gql(queryEnum.value),
+    variables: variables,
+  );
+  //TODO parse exception and throw standard exception and then parse specific errors throw specific exceptions
+  // finally return QueryResult
+  return await graphQLClient.query(_options);
+}
+
+
 enum QueryEnum {
   register,
   login,
